% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dockerfile.R
\name{create_dockerfile}
\alias{create_dockerfile}
\title{create the dockerfile to run the Shiny app without using shiny-server}
\usage{
create_dockerfile(
  deps_path,
  path = ".",
  app_dir = NA,
  r_ver = "latest",
  tlmgr = character(0),
  ncpus = 4,
  minimal = FALSE,
  golem_package_name = NA,
  gh_pat = NA,
  ...
)
}
\arguments{
\item{path}{the path to write the Dockerfile to.}

\item{app_dir}{Directory containing application. If \code{NA}, defaults to the directory
containing the "deps.yaml" file.}

\item{r_ver}{the version of R to use.  Defaults to "latest".}

\item{tlmgr}{character vector of TeX Live packages that should be included in the
installation of dependencies.}

\item{ncpus}{the number of parallel processes to use for a parallel install of
more than one source package. (See \code{Ncpus} argument from \code{utils::install.packages})}

\item{minimal}{Defaults to \code{FALSE}. If \code{TRUE}, strips down the Dockerfile,
removing extra code such as \code{tryCatch}. (\emph{NOTE:} For internal use)}

\item{golem_package_name}{the name of the package as a character string.}

\item{deps_yaml}{the "deps.yaml" file created by \code{automagic::make_deps_file()}}
}
\value{
nothing
}
\description{
This function creates the Dockerfile to run a shiny app without shiny-server.
}
\examples{

\dontrun{



# run the following from the main folder of a Shiny app


  polishedapi::create_dockerfile("deps.yaml", tlmgr = "fancyhdr")

  polishedapi::create_dockerfile("deps.yaml", tlmgr = c(
"fancyhdr",
"multicol",
"sectsty",
"xcolor",
"hyperref",
"booktabs",
"longtable",
"array",
"multirow",
"wrapfig",
"float",
"colortbl",
"pdflscape",
"tabu",
"threeparttable",
"threeparttablex",
"normalem",
"makecell",
"comment",
"amsmath",
"iftex",
"kvoptions",
"etoolbox",
"pdftexcmds",
"infwarerr",
"geometry",
"fancyvrb",
"framed",
"trimspaces",
"ulem",
"epstopdf-pkg"
))
}


}
