% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_packages.R
\name{install_packages}
\alias{install_packages}
\title{install_packages}
\usage{
install_packages(deps_path, minimal, gh_pat, ...)
}
\arguments{
\item{deps_path}{file path to the deps.json file.}

\item{minimal}{T/F}

\item{gh_pat}{GitHub PAT}

\item{...}{additional arguments to pass to the install function.}
}
\description{
install_packages
}
